requirements.txt
cffi==1.15.0
cryptography==35.0.0
Django==2.2
djangorestframework==3.12.4
Pillow==8.4.0
pycparser==2.20
PyJWT==2.3.0
pytz==2021.3
